requirements.txt
numpy
glfw
PyOpenGL
Pillow